+++ packages/target-tauri/src-tauri/src/html_window/mod.rs
@@ -4,8 +4,8 @@ use anyhow::anyhow;
 use log::{error, info, trace, warn};
 
 use tauri::{
-    async_runtime::block_on, webview::WebviewBuilder, LogicalPosition, LogicalSize, Manager, Url,
-    Webview, WebviewUrl, Window, WindowBuilder, WindowEvent,
+    async_runtime::block_on, webview::WebviewBuilder, LogicalPosition, LogicalSize, Manager, State,
+    Url, Webview, WebviewUrl, Window, WindowBuilder, WindowEvent,
 };
 use tauri_plugin_dialog::DialogExt;
 use tauri_plugin_opener::OpenerExt;
@@ -16,13 +16,17 @@ use crate::{
         error::Error,
         punycode::{puny_code_decode_host, puny_code_encode_host},
     },
+    menus::{
+        float_on_top::set_window_float_on_top_based_on_main_window,
+        html_window_menu::create_html_window_menu,
+    },
     settings::{
-        get_content_protection, get_setting_bool_or, CONFIG_FILE,
+        apply_zoom_factor_html_window, get_content_protection, get_setting_bool_or, CONFIG_FILE,
         HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_DEFAULT, HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_KEY,
     },
     state::html_email_instances::InnerHtmlEmailInstanceData,
     temp_file::get_temp_folder_path,
-    DeltaChatAppState, HtmlEmailInstancesState,
+    DeltaChatAppState, HtmlEmailInstancesState, MenuManger,
 };
 
 const HEADER_HEIGHT: f64 = 100.;
@@ -38,8 +42,9 @@ mod punycode;
 #[tauri::command]
 pub(crate) async fn open_html_window(
     app: tauri::AppHandle,
-    html_instances_state: tauri::State<'_, HtmlEmailInstancesState>,
-    dc: tauri::State<'_, DeltaChatAppState>,
+    html_instances_state: State<'_, HtmlEmailInstancesState>,
+    dc: State<'_, DeltaChatAppState>,
+    menu_manager: State<'_, MenuManger>,
     window_id: &str,
     account_id: u32, // TODO needs to be used later for fetching webrequests over dc core
     is_contact_request: bool,
@@ -227,8 +232,7 @@ pub(crate) async fn open_html_window(
 
     let header_view_arc = Arc::new(header_view);
     let mail_view_arc = Arc::new(mail_view);
-    let window_arc = Arc::new(window);
-    let window = window_arc.clone();
+    let window_arc = Arc::new(window.clone());
 
     // resize
     window.on_window_event(move |event| {
@@ -275,6 +279,23 @@ pub(crate) async fn open_html_window(
         truncate_text(subject, 42),
         truncate_text(sender, 40)
     ))?;
+
+    if let Err(err) = apply_zoom_factor_html_window(&app) {
+        error!("failed to apply zoom factor: {err}")
+    }
+    if let Err(err) = set_window_float_on_top_based_on_main_window(&window) {
+        error!("failed to apply float on top: {err}")
+    }
+
+    let window_clone = window.clone();
+    menu_manager
+        .register_window(
+            &app,
+            &window,
+            Box::new(move |app| create_html_window_menu(app, &window_clone)),
+        )
+        .await
+        .map_err(|err| Error::MenuCreation(err.to_string()))?;
 
     Ok(())
 }