+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - add a way to edit messages (in development mode only) #4717
 - tauri: add support for sticker picker #4707
 - tauri: add html email view #4699
-- tauri: add titlebar menu #4755
+- tauri: add titlebar menu #4755 #4787
 - tauri: implement runtime.copyFileToInternalTmpDir and allow opening files in temp dir #4778
 
 ### Changed